Adicionando validações de campos e reformulado respostas de erros.

+++ src/controllers/VariableController.js
@@ -5,6 +5,9 @@ class VariableController {
   async createVariable(req, res) {
     try {
       const { name, type } = req.body;
+if((name==null)||(name==="")||(type==null)||(type<0)){
+  return res.status(400).json({error:"Parâmetros inválidos. Verifique o json enviado."})
+}
 
       const newVariable = await prisma.playerVariable.create({
         data: {
@@ -14,22 +17,22 @@ class VariableController {
       });
 
       return res.status(200).json({
-        code:200,
-        msg:"Sucesso",
         id: newVariable.id,
         name: newVariable.name,
         type: newVariable.type,
       });
     } catch (error) {
       console.error("Ocorreu um erro:", error);
-      return res.status(500).json({code:500,  msg: "Erro interno do servidor" });
+      return res.status(500).json({error: "Erro interno do servidor" });
     }
   }
 
   async updateVariable(req, res) {
     try {
       const { id, name, type } = req.body;
-
+if((name==null)||(name==="")||(type==null)||(type<0)||(id==null)||(id<=0)){
+  return res.status(400).json({error:"Verifique o json enviado. Alguns parâmetros são inválidos"});
+}
       const updatedVariable = await prisma.playerVariable.update({
         where: {
           id,
@@ -41,46 +44,46 @@ class VariableController {
       });
 
       return res.status(200).json({
-        code:200,
-        msg:"Sucesso",
         id: updatedVariable.id,
         name: updatedVariable.name,
         type: updatedVariable.type,
       });
     } catch (error) {
       console.error("Ocorreu um erro:", error);
-      return res.status(500).json({code:200,  msg: "Erro interno do servidor" });
+      return res.status(500).json({error: "Erro interno do servidor" });
     }
   }
 
   async deleteVariable(req, res) {
     try {
       const id = parseInt(req.params.id)
-
+if((id==null)||(id<=0)){
+  return res.status(400).json({error:"Id inválido"})
+}
       await prisma.playerVariable.delete({
         where: {
           id,
         },
       });
 
-      return res.status(200).json({code:200, msg:"Sucesso"});
+      return res.status(200).json({id: id});
     } catch (error) {
       console.error("Ocorreu um erro:", error);
-      return res.status(500).json({code:500,  msg: "Erro interno do servidor" });
+      return res.status(500).json({error: "Erro interno do servidor" });
     }
   }
 
   async getListVariables(req, res) {
     try {
-      let listIds = req.params.listIds
+      let listIds = req.params.listIds;
 
       if (!listIds) {
-        return res.status(400).json({ message: "IDs não informados" });
+        return res.status(400).json({ error: "IDs não informados" });
       }
 
-      listIds = listIds.split(",").map(id => parseInt(id, 10))
+      listIds = listIds.split(",").map(id => parseInt(id, 10));
 
-      const variable = await prisma.playerVariable.findMany({
+      const variables = await prisma.playerVariable.findMany({
         where: {
           id: {
             in: listIds
@@ -88,22 +91,25 @@ class VariableController {
         },
       });
 
-      if (!variable) {
-        return res.status(404).json({code:404,  msg: "Variável não encontrada" });
+      if (!variables || variables.length === 0) {
+        return res.status(404).json({error: "Variáveis não encontradas" });
       }
 
-      return res.status(200).json({
-        code:200,
-        msg:"Sucesso",
+      const variableList = variables.map(variable => ({
         id: variable.id,
         name: variable.name,
         type: variable.type,
-      });
+      }));
+
+      return res.status(200).json(
+        variableList,
+      );
     } catch (error) {
       console.error("Ocorreu um erro:", error);
-      return res.status(500).json({code:500,  msg: "Erro interno do servidor" });
+      return res.status(500).json({error: "Erro interno do servidor" });
     }
   }
+
 }
 
 export default VariableController;
